+++ src/main/java/com/example/swp/entities/OrderDetails.java
@@ -16,7 +16,7 @@ public class OrderDetails {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    private Long quantity;
+    private int quantity;
 
     @JsonProperty(value = "unit_price")
     private double unitPrice;
